requirements.txt
fastapi==0.111.0
python-multipart==0.0.27
uvicorn==0.29.0
streamlit==1.35.0
openai==1.30.1
httpx==0.27.2
python-dotenv==1.0.1
tiktoken==0.7.0
hypothesis==6.103.1
pytest==8.2.1
requests==2.32.3
Pillow==10.3.0
beautifulsoup4==4.12.3
